Added error handling for voicemeeter check to vm plugin.

+++ api/voicemeeter.py
@@ -12,9 +12,11 @@ class status:
 
 def main():
     while True:
-        if ("voicemeeter8.exe" in (p.name() for p in psutil.process_iter())) == False:
-            os.system("taskkill /f /im voicemeeter8.exe")
-            os.system('start /b "" ..\..\Voicemeeter.lnk')
+        if os.path.exists("C:\Program Files (x86)\VB\Voicemeeter\\voicemeeter8.exe"):
+            if os.path.exists("..\..\Voicemeeter.lnk"):
+                if ("voicemeeter8.exe" in (p.name() for p in psutil.process_iter())) == False:
+                    os.system("taskkill /f /im voicemeeter8.exe")
+                    os.system('start /b "" ..\..\Voicemeeter.lnk')
         time.sleep(5)
         status.vars['lastLoop'] = pytools.clock.getDateTime()
         status.finishedLoop = True
